--- a/bench.mjs
+++ b/bench.mjs
@@ -1,0 +1,10 @@
+const pattern = new URLPattern({ pathname: "/" });
+
+console.log(pattern.exec({ pathname: "/" }));
+
+Deno.bench({
+  name: "URLPattern",
+  fn() {
+    pattern.exec({ pathname: "/" });
+  },
+});