Currently a POC. Rewrite `urlpattern::quirks` in JS to avoid Rust<->JS trips.

2.5x faster in simple pathname matching.

```
benchmark       time (avg)        iter/s             (min … max)       p75       p99      p995
---------------------------------------------------------------- -----------------------------
URLPattern      3.11 µs/iter      321,443.2      (3.05 µs … 3.49 µs)    3.11 µs   3.49 µs   3.49 µs
URLPattern      1.24 µs/iter     803,437.8      (1.23 µs … 1.3 µs)       1.25 µs    1.3 µs    1.3 µs
```